Added comments and separated document retrieval functions into new file
